% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analizar_bivariados.R
\name{analizar_bivariados}
\alias{analizar_bivariados}
\title{Analizar bivariados Function}
\usage{
analizar_bivariados(
  data,
  variable,
  var_obj = "RESULTADO",
  default_levels = 4,
  descriptivos = TRUE,
  p_value_threshold = 0.1,
  max_levels = 10,
  robust = TRUE,
  plot = TRUE
)
}
\arguments{
\item{data}{Dataset selected}

\item{variable}{Variable to analize}

\item{var_obj}{RESULTADO Variable}

\item{default_levels}{Default = 4}

\item{descriptivos}{Default = TRUE}

\item{p_value_threshold}{Default = 0.1}

\item{max_levels}{Default = 10}

\item{robust}{Default = TRUE}

\item{plot}{Default = TRUE}
}
\value{
results
}
\description{
Analizar bivariados Function
}
