% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_scores_logit.R
\name{calcular_scores_logit}
\alias{calcular_scores_logit}
\title{Calcular Scores Logit}
\usage{
calcular_scores_logit(id_tibble, logit, especificacion_modelo, bd_madre)
}
\arguments{
\item{id_tibble}{ID}

\item{logit}{Model Result}

\item{especificacion_modelo}{Model Formula}

\item{bd_madre}{Main Dataset}
}
\value{
score
}
\description{
Calcular Scores Logit
}
