% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculoGini.R
\name{calculoGini}
\alias{calculoGini}
\title{calculoGini Function}
\usage{
calculoGini(newData, modelo, VarDep = "RESULTADO")
}
\arguments{
\item{newData}{DataBase}

\item{modelo}{Estimation output}

\item{VarDep}{Default = RESULTADO}
}
\value{
res_perf
}
\description{
calculoGini Function
}
