% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculoGini2.R
\name{calculoGini2}
\alias{calculoGini2}
\title{CalculoGini2 Function}
\usage{
calculoGini2(VarEstimada, res_perf, base2, VarDep)
}
\arguments{
\item{VarEstimada}{Estimated variable}

\item{res_perf}{Performance res-}

\item{base2}{DataBase}

\item{VarDep}{Endogenous variable from base2}
}
\value{
res_perf
}
\description{
CalculoGini2 Function
}
