% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters_to_factors.R
\name{characters_to_factor}
\alias{characters_to_factor}
\title{Characters to factor Function}
\usage{
characters_to_factor(dataset, not_factor = NULL)
}
\arguments{
\item{dataset}{Dataset}

\item{not_factor}{Default = NULL}
}
\value{
dataset
}
\description{
Characters to factor Function
}
