% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementar_info_anterior.R
\name{complementar_info_anterior}
\alias{complementar_info_anterior}
\title{Title}
\usage{
complementar_info_anterior(
  dataset_original,
  dataset_adicional,
  cols_to_complete,
  id_cols = c("ctotal_idinstancia"),
  ordering_col = "id_current_stage"
)
}
\arguments{
\item{ordering_col}{}
}
\description{
Title
}
