% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completar_character_columns.R
\name{completar_character_columns}
\alias{completar_character_columns}
\title{Completar character columns Function}
\usage{
completar_character_columns(dataset, cutoff_nas = 0.8, not_complete = NULL)
}
\arguments{
\item{dataset}{Dataset}

\item{cutoff_nas}{Default = 0.8}

\item{not_complete}{Default = NULL}
}
\value{
dataset
}
\description{
Completar character columns Function
}
