% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion_iv_robusta.R
\name{discretizacion_iv_robusta}
\alias{discretizacion_iv_robusta}
\title{Discretizacion IV robusta Function}
\usage{
discretizacion_iv_robusta(
  vector_variable,
  num_levels,
  intentos_por_num_levels = 10
)
}
\arguments{
\item{vector_variable}{Original Variable/vector}

\item{num_levels}{Number of Levels}

\item{intentos_por_num_levels}{Default = 10}
}
\value{
vector_discretizado
}
\description{
Discretizacion IV robusta Function
}
