% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizar.R
\name{discretizar}
\alias{discretizar}
\title{discretizar function}
\usage{
discretizar(
  vector_valores,
  puntos_corte = c(-Inf, 2, 10, 30, Inf),
  etiquetas = c("unpredictive", "weak", "medium", "strong")
)
}
\arguments{
\item{etiquetas}{}
}
\description{
discretizar function
}
