% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_axis_graph.R
\name{double_axis_graph}
\alias{double_axis_graph}
\title{Double Axis Graph Function}
\usage{
double_axis_graph(graf1, graf2)
}
\arguments{
\item{graf1}{Plot n 1}

\item{graf2}{Plot n 2}
}
\value{
graf
}
\description{
Return the merge of two plots into a single one
}
