% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluar_felicidad_variables.R
\name{evaluar_felicidad_variables}
\alias{evaluar_felicidad_variables}
\title{Evaluar Felicidad Variables Function}
\usage{
evaluar_felicidad_variables(id_tibble, logit, especificacion_modelo, bd_madre)
}
\arguments{
\item{id_tibble}{ID}

\item{logit}{Logit model}

\item{especificacion_modelo}{Model Formula}

\item{bd_madre}{Main DataBase}
}
\value{
tibble_evaluacion
}
\description{
Evaluar Felicidad Variables Function
}
