% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_categoricas.R
\name{extraer_categoricas}
\alias{extraer_categoricas}
\title{Extraer Categoricas Function}
\usage{
extraer_categoricas(lista_seleccion, lista_variables)
}
\arguments{
\item{lista_seleccion}{Globl list of variables}

\item{lista_variables}{List of variables}
}
\value{
categoricas_obtenidas
}
\description{
Extraer Categoricas Function
}
