% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtro1obs.R
\name{Filtro1obs}
\alias{Filtro1obs}
\title{Filtro1obs Function}
\usage{
Filtro1obs(BDModelo)
}
\arguments{
\item{BDModelo}{Uncleaned DataBase}
}
\value{
BDModelo
}
\description{
Exports a csv table that resumes observations per variable and
returns BDModelo with variables more than 1 observations
}
