% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiltroMissing.R
\name{FiltroMissing}
\alias{FiltroMissing}
\title{FiltroMissing Function}
\usage{
FiltroMissing(BDModelo, p, nmissxvar)
}
\arguments{
\item{BDModelo}{Uncleaned DataBase}

\item{p}{Threshold of minimimun missing observations}

\item{nmissxvar}{Name of the filtered Data Base}
}
\value{
nmissxvar
}
\description{
Returns the number of missings per variable
}
