% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complete_variables.R
\name{get_complete_variables}
\alias{get_complete_variables}
\title{Get complete variables Function}
\usage{
get_complete_variables(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
dataset
}
\description{
Get complete variables Function
}
