% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPicturesAttrition.R
\name{getPicturesAttrition}
\alias{getPicturesAttrition}
\title{getPicturesAttrition Function}
\usage{
getPicturesAttrition(
  attritionBase,
  varResponseRaw,
  varNamesModelRaw,
  varNamesModelLenRaw,
  nameFileRaw
)
}
\arguments{
\item{nameFileRaw}{}
}
\value{

}
\description{
getPicturesAttrition Function
}
