% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHH.R
\name{IHH}
\alias{IHH}
\title{IHH Function}
\usage{
IHH(x, num_elem = NULL)
}
\arguments{
\item{x}{Vector}

\item{num_elem}{Set to true if there are Null values in vector. Default = NULL}
}
\value{
ihh
}
\description{
IHH Function
}
