% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputacionVar.R
\name{ImputacionVar}
\alias{ImputacionVar}
\title{ImputacionVar Function}
\usage{
ImputacionVar(BDModelo)
}
\arguments{
\item{BDModelo}{Uncleaned Dataset}
}
\value{
BDModelo
}
\description{
Funcion imputacion de variables:
- numericas con el promedio
- categoricas con la categoria de mayor frecuencia.
}
