% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVWOE.R
\name{IVWOE}
\alias{IVWOE}
\title{IVWOE Function}
\usage{
IVWOE(Btrain, varNames, VarRpta)
}
\arguments{
\item{Btrain}{Train DataSet}

\item{varNames}{Variable Names from Train Dataset}

\item{VarRpta}{Variable Names}
}
\value{
IV
}
\description{
IVWOE Function
}
