% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVxVar.R
\name{IVxVar}
\alias{IVxVar}
\title{IVxVar Function}
\usage{
IVxVar(Btrain, varNames, varResponse = "RESULTADO")
}
\arguments{
\item{Btrain}{train DataSet}

\item{varNames}{Variable Names from Train Dataset}

\item{varResponse}{Variable Names}
}
\description{
IVxVar Function
}
