% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo_lasso.R
\name{modelo_lasso}
\alias{modelo_lasso}
\title{Modelo lasso}
\usage{
modelo_lasso(
  dataset,
  woe = FALSE,
  graficar = TRUE,
  titulo = "modelo lasso",
  ridge = FALSE
)
}
\arguments{
\item{dataset}{Dataset}

\item{woe}{Default = TRUE}

\item{graficar}{Default = TRUE}

\item{titulo}{Default = "modelo lasso"}

\item{ridge}{Default = FALSE}
}
\value{
results
}
\description{
Modelo lasso
}
