% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo_lasso_parsimonioso.R
\name{modelo_lasso_parsimonioso}
\alias{modelo_lasso_parsimonioso}
\title{Modelo lasso parsimonioso Function}
\usage{
modelo_lasso_parsimonioso(
  dataset,
  max_variables = 20,
  woe = FALSE,
  graficar = TRUE,
  titulo = "modelo lasso parsimonioso"
)
}
\arguments{
\item{titulo}{}
}
\description{
Modelo lasso parsimonioso Function
}
