% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo_logit.R
\name{modelo_logit}
\alias{modelo_logit}
\title{Modelo Logit Function}
\usage{
modelo_logit(
  dataset,
  covariables,
  target_variable = "RESULTADO",
  woe = FALSE,
  graficar = TRUE,
  titulo = "modelo logit"
)
}
\arguments{
\item{titulo}{}
}
\value{

}
\description{
Modelo Logit Function
}
