% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomalizar_variable.R
\name{nomalizar_variable}
\alias{nomalizar_variable}
\title{Nomalizar variable Function}
\usage{
nomalizar_variable(variable_vector, critical_value = 1.96)
}
\arguments{
\item{critical_value}{}
}
\value{

}
\description{
Nomalizar variable Function
}
