% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediccion_arbol.R
\name{prediccion_arbol}
\alias{prediccion_arbol}
\title{Prediccion Arbol Function}
\usage{
prediccion_arbol(
  df,
  modelo_1,
  modelo_2 = NULL,
  target = "ESTADO_FINAL",
  destination = "FLAG_PREDICCION_ARBOL",
  var_skip = NULL,
  id_variables = c("ID")
)
}
\arguments{
\item{id_variables}{}
}
\value{

}
\description{
Prediccion Arbol Function
}
