% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare dataset Function}
\usage{
prepare_dataset(dataset, not_numeric = NULL, not_factor = NULL)
}
\arguments{
\item{not_factor}{}
}
\value{

}
\description{
Prepare dataset Function
}
