% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitar.R
\name{quitar}
\alias{quitar}
\title{Quitar Function}
\usage{
quitar(vector_original, vector_remover)
}
\arguments{
\item{vector_original}{input vector}

\item{vector_remover}{vector to remove from input}
}
\value{
resultado
}
\description{
Quitar Function
}
\examples{
quitar(c(1:10),c(1:5))
quitar(c("a","b","c"),"a")
}
