% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_constant_factors.R
\name{remove_constant_factors}
\alias{remove_constant_factors}
\title{Remove constant factor Function}
\usage{
remove_constant_factors(dataset, default_dummy = 0)
}
\arguments{
\item{default_dummy}{}
}
\description{
Remove constant factor Function
}
