% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumir_modelo_logit.R
\name{resumir_modelo_logit}
\alias{resumir_modelo_logit}
\title{Resumir modelo logit Function}
\usage{
resumir_modelo_logit(logit, nombre_modelo)
}
\arguments{
\item{nombre_modelo}{}
}
\description{
Resumir modelo logit Function
}
