% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumir_variables.R
\name{resumir_variables}
\alias{resumir_variables}
\title{Resumir variables Function}
\usage{
resumir_variables(dataset, excepciones_remover = NULL, missing_threshold = 0.1)
}
\arguments{
\item{excepciones_remover}{Default = NULL}

\item{missing_threshold}{Default = 0.1}
}
\description{
Resumir variables Function
}
