% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncar_outliers.R
\name{truncar_outliers}
\alias{truncar_outliers}
\title{Truncar outliers Function}
\usage{
truncar_outliers(
  dataset,
  percentiles = c(0.01, 0.09),
  porc_tolerancia_negativos = 0.01
)
}
\arguments{
\item{percentiles}{Default = c(0.001,0.009)}

\item{porc_tolerancia_negativos}{Default = 0.01}
}
\description{
Truncar outliers Function
}
