% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncar_variable.R
\name{truncar_variable}
\alias{truncar_variable}
\title{Truncar variable Function}
\usage{
truncar_variable(
  dataset,
  nombre_variable,
  percentiles = c(0.01, 0.09),
  porc_tolerancia_negativos = 0.01
)
}
\arguments{
\item{percentiles}{Default = c(0.01,0.09)}

\item{porc_tolerancia_negativos}{Default= 0.01}
}
\description{
Truncar variable Function
}
