% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_to_numeric.R
\name{try_to_numeric}
\alias{try_to_numeric}
\title{Title}
\usage{
try_to_numeric(x, not_numeric = NULL)
}
\arguments{
\item{x}{Vector}

\item{not_numeric}{List of excepted variable}
}
\value{

}
\description{
Title
}
