% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_naa_data.R
\name{validate_naa_data}
\alias{validate_naa_data}
\title{Validate data for Dashboard}
\usage{
validate_naa_data(df, age_classes, ndraws = 1, years = 1)
}
\arguments{
\item{df}{long dataframe with numbers at age.}

\item{age_classes}{number of age classes.}

\item{ndraws=1}{number of random draws in the dataframe}

\item{years=1}{number of years in the dataframe}
}
\description{
Validation function that ensure that the data we have is the
data we want to push to the dashboard
}
\details{
df is expcted to contain character variables for
fishery, common name, metric, source,  units.
df is expected to contain numerics for
itis code (species_itis) and value.
df is expectd to contain a data_version, which is a Date
df is expected to contain state and wave. These may be NA
ndraws and years are used with age_classes to ensure the proper number of rows
#' @return nothing

@export
}
