% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOExVar.R
\name{WOExVar}
\alias{WOExVar}
\title{WOExVar Function}
\usage{
WOExVar(Btrain, varNames, varResponse = "RESULTADO")
}
\arguments{
\item{Btrain}{Train DataSet}

\item{varNames}{Variable Names from Train Dataset}

\item{varResponse}{Default = RESULTADO}
}
\value{
Btrain
}
\description{
WOExVar Function
}
