% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer_setup.R
\name{developer_setup}
\alias{developer_setup}
\title{Set up and Create, if necessary, Paths and Directories for your project}
\usage{
developer_setup(dev_name, fishery, management_year)
}
\arguments{
\item{dev_name}{Developer Initials.}

\item{fishery}{Either flukeRDM or groundfishRDM.}

\item{management_year}{the fishing year in which fishery management actions will take place}
}
\value{
A file path to the data storage directory.
}
\description{
Sets up paths for various developers.
}
\details{
Not much to say here
}
\examples{
developer_setup("mlee","groundfishRDM", 2028)#'
#> "C:/Users/min-yang.lee/Documents/dstoolkit/Data/2028_mgt_cycle"
}
