% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizar.R
\name{discretizar}
\alias{discretizar}
\title{Discretizar Function}
\usage{
discretizar(
  vector_valores,
  puntos_corte = c(-Inf, 2, 10, 30, Inf),
  etiquetas = c("unpredictive", "weak", "medium", "strong")
)
}
\arguments{
\item{vector_valores}{Original Variable/vector}

\item{puntos_corte}{Default = c(-Inf,2,10,30,Inf)}

\item{etiquetas}{Default = c("unpredictive","weak","medium","strong")}
}
\value{
vector_clasificacion
}
\description{
Discretizar Function
}
