% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_naa_long.R
\name{pivot_naa_long}
\alias{pivot_naa_long}
\title{Reshape Numbers at Age to Long}
\usage{
pivot_naa_long(df)
}
\arguments{
\item{df}{dataframe with columns containing numbers at age, those columns
start with "age"}
}
\value{
A long dataframe
}
\description{
Data dashboard helper file that reshapes NAA from Wide to Long format
}
\details{
A simple grep and pivot longer
}
\examples{
sample_df <- tibble(
  year   = c(2001:2002),
  metric = rep(c("NAA"), times=2),
  age2   = c(500, 150),
  age3   = c(250, 100))
> pivot_naa_long(sample_df)
 # A tibble: 4 × 3
 year metric value
 <int> <glue> <dbl>
 1  2001 NAA 2    500
 2  2001 NAA 3    250
 3  2002 NAA 2    150
 4  2002 NAA 3    100#'
}
